+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # HEAD
 
+-   **Breaking change** New API for noise module [#151](https://github.com/litebird/litebird_sim/pull/151):
+
+    -   Function `add_noise` has been renamed to `add_noise_to_observations`, and its parameter `noisetype` has been renamed into `noise_type` for consistency with other parameters (**breaking**)
+
+    -   New functions `add_white_noise` and `add_one_over_f_noise` are exported (they were already implemented but were not visible)
+
+    -   Each `Simulation` object creates random number generators (field `Simulation.random`), in a way that is safe even for MPI applications
+
 -   Upgrade NumPy from 1.20 to 1.21, Numba from 0.54 to 0.55, Rich from 6.2 to 11.0 [#152](https://github.com/litebird/litebird_sim/pull/152)
 
 -   Fix issue [#148](https://github.com/litebird/litebird_sim/issues/148)